% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_coefs}
\alias{extract_coefs}
\title{Extract coefficients from a fitted model and give the coefficients more informative names based on the specified formulas.}
\usage{
extract_coefs(mod, CV = FALSE, size0 = FALSE)
}
\arguments{
\item{mod}{an object of type \code{growmod} that was fit using the \code{growmod()} function. See \code{?growmod} for details.}

\item{CV}{logical - Should estimates of the coefficients of variance (variance/intercept) be reported?}

\item{size0}{logical - Should estimates of the distribution of initial sizes be reported?}
}
\description{
Extract coefficients from a fitted model and give the coefficients more informative names based on the specified formulas.
}

