% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_ID_dev}
\alias{extract_ID_dev}
\title{Extract estimates of individual deviates from the average pattern.}
\usage{
extract_ID_dev(mod, data)
}
\arguments{
\item{mod}{an object of type \code{growmod} that was fit to \code{data} using the \code{growmod()} function. See \code{?growmod} for details.}

\item{data}{the same data set that was used for fitting \code{mod}}
}
\value{
a data frame with the estimated individual deviations from the average intercept added to the original observations and covariates. The data fram has repeated observations of the same individual at the same point averaged out.
}
\description{
Extract estimates of individual deviates from the average pattern.
}

