% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_pred}
\alias{extract_pred}
\title{Extract estimates of the latent variable}
\usage{
extract_pred(mod, data)
}
\arguments{
\item{mod}{an object of type \code{growmod} that was fit to \code{data} using the \code{growmod()} function. See \code{?growmod} for details.}

\item{data}{the same data set that was used for fitting \code{mod}}
}
\value{
predictions from a growmod model on the same scale as the sizes provided (e.g. log kg)
}
\description{
Extract estimates of the latent variable
}

