% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{interpolate}
\alias{interpolate}
\title{Extract estimates of the latent variable}
\usage{
interpolate(mod, data)
}
\arguments{
\item{mod}{an object of type \code{growmod} that was fit to \code{data} using the \code{growmod()} function. See \code{?growmod} for details.}

\item{data}{the same data set that was used for fitting \code{mod}}
}
\value{
A data set containing interpolated values \code{size.interpolated}. These come from the estimated latent states so that, in addition to the fixed effects, these include process noise and any random effects.
}
\description{
Extract estimates of the latent variable
}

