% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{LRtest}
\alias{LRtest}
\title{Likelihood Ratio Test}
\usage{
LRtest(full, restricted)
}
\arguments{
\item{full}{an object of type \code{growmod} that was fit using the \code{growmod()} function.}

\item{restricted}{an object of type \code{growmod} that was fit using the \code{growmod()} function with a subset of the covariates used for \code{full}}
}
\description{
Likelihood Ratio Test
}

