% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{organize_data}
\alias{organize_data}
\title{Organize data so that it can be read into the TMB model}
\usage{
organize_data(data, sigma_obs)
}
\arguments{
\item{data}{a data frame of observed sizes and predictors in long format. See details.}

\item{sigma_obs}{the standard deviation of the observation error with the same units as size (e.g. log kg). To be used in a \code{growmod} model when \code{estobserr=FALSE}.}
}
\value{
a list of objects in the correct format to be read in by the TMB model defined in \code{growmod.cpp} with the exception of beta, eta, M, and X which depend on the model formulas and are defined in the growmod function.
}
\description{
Organize data so that it can be read into the TMB model
}
\details{
data must contain at least the following columns: \code{size, t, t0, and ID}. \code{size} is the size of an individual on the log scale. \code{t} is the point in time when an observation was made. \code{t} must be in discrete units (i.e. integers) with increments of 1. \code{t0} is the time when an individual was born. \code{ID} is a unique identifier for each individaul. It must contain at least one row per combination of individual and time. Multiple rows may be needed if an individual was measured more than once at the same timepoint. These multiple measurements are very informative for the estimation of observation error. If multiple rows for the same combination of individual and time are given, then only the covariates from the first instance is used. If a size observation was missed for one combination of individual and time, then that row should have an \code{NA} in the size column, but must contian any covariates. There must be at least one non-missing size for each individual.
}

