% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simobs}
\alias{simobs}
\title{Simulate observed sizes.}
\usage{
simobs(pars, nind = 150, ntime = 29, maxage = 12, lifespans = NULL,
  recap = 0.5, sigma_obs = 0.05)
}
\arguments{
\item{pars}{a numeric vector contianing the following named terms representing regression coefficients: \code{'(Intercept)', 'age',  'I(age^2)', 'prevsize', 'prevsize:age', 'sigma_proc', 'time_growth_sd', 'indiv_growth_sd', 'indiv_age_growth_sd', 'indiv_cor', 'size0_mu', 'size0_sd'}}

\item{nind}{the number of individuals to simulate}

\item{ntime}{the number of times spanned by the dataset}

\item{maxage}{the maximum age that all individuals reach (not used if lifespans is specified)}

\item{lifespans}{a set of intergers to be resampled from. It should reflect the frequency of each lifespan. The units are the same as for times.}

\item{recap}{the probability of captuing an individual in any time when it was alive.}

\item{sigma_obs}{the standard deviation of the observation error with the same units as size (e.g. log kg).}
}
\value{
A data frame of the format needed for input to \code{growmod}. Size observations that are missing due to imperfect recapture are entered as NA because rows of predictors are still needed for those combinations of individual and time.
}
\description{
Simulate observed sizes.
}
\details{
Covariates other than age are not implemented yet. This function calls \code{simgrow()}. It assumes that there are no reapeated captures of the same individuaal at the same timepoint (less than ideal for estimating sigma_obs).
}
\examples{
true=c("(Intercept)"=1, age=0,  "I(age^2)"=0, size=.6, "size:age"=0, sigma_proc=.01, time_growth_sd=.1, indiv_growth_sd=.1, indiv_age_growth_sd=0, indiv_cor=0, size0_mu=2.5, size0_sd=1)
sg=simobs(true, nind=10, ntime=20)
}

