% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.growmod}
\alias{summary.growmod}
\title{Output estimates and summary statistics from a sucessfully fitted model}
\usage{
\method{summary}{growmod}(mod)
}
\arguments{
\item{mod}{an object of type \code{growmod} that was fit using the \code{growmod()} function. See \code{?growmod} for details.}
}
\description{
return summary statistics for a model fitted by the \code{growmod()} function.
}

