% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TMBBIC}
\alias{TMBBIC}
\title{Calculate the Bayesian Information Criteria}
\usage{
TMBBIC(opt, n = NULL)
}
\arguments{
\item{opt}{an optimized TMB model}

\item{n}{number of observations}
}
\description{
Calculate the Bayesian Information Criteria
}
