% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simgrow}
\alias{simgrow}
\title{Simulate growth}
\usage{
simgrow(pars, nind = 150, ntime = 29, maxage = 12, lifespans = NULL)
}
\arguments{
\item{pars}{a numeric vector contianing the following named terms representing regression coefficients: \code{'(Intercept)', 'age',  'I(age^2)', 'size', 'size:age', 'sigma_proc', 'time_growth_sd', 'indiv_growth_sd', 'indiv_age_growth_sd', 'indiv_cor', 'size0_mu', 'size0_sd'}}

\item{nind}{the number of individuals to simulate}

\item{ntime}{the number of times spanned bt the dataset}

\item{maxage}{the maximum age that all individuals reach (not used if lifespans is specified)}

\item{lifespans}{a set of intergers to be resampled from. It should reflect the frequency of each lifespan.}
}
\description{
Simulate growth
}
\details{
Covariates other than age are not implemented. Birth times are uniformly distributed across the time series. Observations of an individual stop either when it reaches the end of its lifespan, or when time reaches \code{ntime}.
}

